backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
requests
pillow
google-generativeai
google-cloud-translate
openai
